pyproject.toml
[project]
name = "tmdb-cli-tool"
version = "0.1.0"
description = ""
authors = [
    {name = "aketrari",email = "anastasiaketrari@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "tabulate (>=0.9.0,<0.10.0)"
]

[tool.poetry]
name = "tmdb-cli-tool"
version = "0.1.0"
description = "A command-line tool for browsing movies from TMDB"
packages = [{include = "tmdb_cli_tool", from = "src"},{ include = "tmdb_cli" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"

[tool.poetry.scripts]
tmdb-app = "main:main"